What the copies search costs in memory, measured (#101)
st4 -c on a 41 KB file ends in OutOfMemoryError under the Java tools at the
default heap, where the Go tools pack it. research.md records where the
memory goes, and two readings that are wrong.

WHAT IT IS NOT. Not retention: after the search returns the heap has 1 MB.
Not the pool's bound: lowering it from four times a full parse to one makes
the peak worse, 6,596 MB against 6,212, and at four the bound is never
reached at all, 0 compactions over the opening passes. Not the width of a
node, though Go kept five fields in int where Java and C# spend four: the
narrowing leaves the peak inside the noise of repeated runs.

WHAT IT IS. A parse makes a node a state reached, and a state is a position
and an offset. One parse of 41,743 units at a window of 32,512 makes
85,971,635 nodes and the opening passes reach 261,766,765, 6,270 a position.
The pool grows geometrically to fit them and every growth abandons the array
it came from: 33 GB allocated in total, 98.2 per cent of it in newNode, for
a live set of about 5 GB. Java's cap is a quarter of the machine, 4 GB of
16, so it is the first to stop.

THE WINDOW IS THE CONTROL, and every asset this format is written for names
one: 6,337 MB at the default 32,512, 2,313 at 960, 610 at 256. A tone-period
column of 19,500 bytes at -m256 costs 75 MB and a tenth of a second.

WHAT CHANGES HERE. The Go pool is int32, as the other two are, which leaves
the bytes out identical and takes the run from 61.0 seconds to 24.7. And the
Java tool reports the heap it ran out of, the units and the window it ran
out at, and the two ways on, rather than ending in a stack trace.

The representation is what costs, a node a state against a winner a
position, and the two parsers beside it show the second works. Bringing the
copies search to it is a redesign of the parse, the checkpoints and the
rebuild together, held to the same bytes. research.md says so and it is not
written here.

+++ go/st4/literalcopy.go
@@ -557,13 +557,18 @@ type copyParser struct {
 	bestMatchIdx   int
 	bestLengthSize int
 
-	// The node pool.
+	// The node pool. The five fields beside the kind are int32, as the
+	// Java and C# pools are: a parse of a long input at a wide window makes
+	// tens of millions of nodes, and Go's int is eight bytes where those
+	// trees spend four. Every value fits: an index and a length are bounded
+	// by the unit count, an offset by the window, and the bits by the
+	// output a 32-bit count already bounds.
 	nodeKind   []byte
-	nodeEnd    []int
-	nodeOffset []int
-	nodeAux    []int
-	nodePred   []int
-	nodeBits   []int
+	nodeEnd    []int32
+	nodeOffset []int32
+	nodeAux    []int32
+	nodePred   []int32
+	nodeBits   []int32
 	nodes      int
 
 	// The literal channel: a min-tree by match end + 1 over bits -
@@ -619,11 +624,11 @@ func newCopyParser(units []uint32, unit, window int) *copyParser {
 		repable:       make([]int, size),
 		inRepable:     make([]bool, size),
 		nodeKind:      make([]byte, 0, 1024),
-		nodeEnd:       make([]int, 0, 1024),
-		nodeOffset:    make([]int, 0, 1024),
-		nodeAux:       make([]int, 0, 1024),
-		nodePred:      make([]int, 0, 1024),
-		nodeBits:      make([]int, 0, 1024),
+		nodeEnd:       make([]int32, 0, 1024),
+		nodeOffset:    make([]int32, 0, 1024),
+		nodeAux:       make([]int32, 0, 1024),
+		nodePred:      make([]int32, 0, 1024),
+		nodeBits:      make([]int32, 0, 1024),
 	}
 	last := map[uint64]int{}
 	for q := 0; q+1 < count; q++ {
@@ -1143,28 +1148,28 @@ func (p *copyParser) newNode(kind byte, end, offset, aux, pred, bits int) int {
 		p.nodeBits = append(p.nodeBits, 0)
 	}
 	p.nodeKind[p.nodes] = kind
-	p.nodeEnd[p.nodes] = end
-	p.nodeOffset[p.nodes] = offset
-	p.nodeAux[p.nodes] = aux
-	p.nodePred[p.nodes] = pred
-	p.nodeBits[p.nodes] = bits
+	p.nodeEnd[p.nodes] = int32(end)
+	p.nodeOffset[p.nodes] = int32(offset)
+	p.nodeAux[p.nodes] = int32(aux)
+	p.nodePred[p.nodes] = int32(pred)
+	p.nodeBits[p.nodes] = int32(bits)
 	p.nodes++
 	return p.nodes - 1
 }
 
 func (p *copyParser) rebuild(last int) *Block {
 	var order []int
-	for node := last; node >= 0; node = p.nodePred[node] {
+	for node := last; node >= 0; node = int(p.nodePred[node]) {
 		order = append(order, node)
 	}
 	chain := &Block{Bits: -1, Index: -1, Offset: InitialOffset}
 	for i := len(order) - 2; i >= 0; i-- {
 		node := order[i]
-		offset := p.nodeOffset[node]
+		offset := int(p.nodeOffset[node])
 		if p.nodeKind[node] == ckLiterals {
 			offset = 0
 		}
-		chain = &Block{Bits: p.nodeBits[node], Index: p.nodeEnd[node],
+		chain = &Block{Bits: int(p.nodeBits[node]), Index: int(p.nodeEnd[node]),
 			Offset: offset, Chain: chain}
 	}
 	return chain
